+++ src/main.ts
@@ -3,6 +3,7 @@ import GLib from 'gi://GLib';
 import * as Utils from './utils.js';
 import app from './app.js';
 import client from './client.js';
+import './overrides.js';
 
 const BIN_NAME = pkg.name.split('.').pop() as string;
 const APP_BUS = (name: string) => `${pkg.name}.${name}`;
