+++ src/service/hyprland.ts
@@ -2,7 +2,6 @@ import GLib from 'gi://GLib';
 import Gio from 'gi://Gio';
 import Service from '../service.js';
 
-Gio._promisify(Gio.DataInputStream.prototype, 'read_upto_async');
 const HIS = GLib.getenv('HYPRLAND_INSTANCE_SIGNATURE');
 
 const socket = (path: string) => new Gio.SocketClient()
