+++ src/utils.ts
@@ -251,23 +251,3 @@ export function subprocess(
     }
 }
 
-// @ts-ignore
-GObject.Object.prototype.toJSON = function() {
-    const result = {};
-    const filter = ['parent', 'window', 'font-options'];
-    const props = (this.constructor as unknown as GObject.ObjectClass)
-        //@ts-ignore
-        .list_properties()
-        .filter(p => !filter.includes(p.name || ''));
-
-    props.forEach(p => {
-        try {
-            //@ts-ignore
-            result[p.name] = this[p.name];
-        }
-        catch (e) {
-            logError(e);
-        }
-    });
-    return result;
-};
